% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_pyramid.R
\name{age_pyramid}
\alias{age_pyramid}
\title{Age-pyramid plot function}
\usage{
age_pyramid(df, age_var = "Alder", gender_var = "Kon", man_level = "Man",
  age_breaks = c(0, 39, 44, 49, 54, 59, 64, 69, 74, 79, 84, Inf),
  age_labels = c("0-39", "40-44", "45-49", "50-54", "55-59", "60-64", "65-69",
  "70-74", "75-79", "80-84", "85+"), percent = TRUE, x_breaks = 5,
  x_breaks_end = 1e+05, title = "", subtitle = NULL, title_size = 14,
  subtitle_size = 10, title_margin = 1, y_lab = NULL, x_lab = NULL,
  background_color = "moccasin", panel_grid_color = "grey",
  panel_grid_size = 0.3, contour_line_size = 0.1, axis_size = 0.3,
  axis_text_angle = 0, text_size = 8, windows_font = "Calibri",
  fill_colors = c("#377EB8", "#E41A1C"), legend_pos = c(0, 0),
  legend_labels = waiver(), label_breaks = waiver(), legend_row = NULL,
  legend_col = NULL)
}
\arguments{
\item{df}{Data frame.}

\item{age_var}{Name of age variable.}

\item{gender_var}{Name of gender variable.}

\item{man_level}{Name of man level, probably "Man" or "Män".}

\item{age_breaks}{Each age observation in an interval.}

\item{age_labels}{Label of the interval.}

\item{percent}{If TRUE, x-axis is in percent form. Otherwise in count form.}

\item{x_breaks}{Length between each break on x-axis.}

\item{x_breaks_end}{Break end, default for 100000. Works for all count values below that.}

\item{title}{Plot title, NULL for no title.}

\item{subtitle}{Small text under title, NULL for no subtitle.}

\item{title_size}{Text size of title in pt.}

\item{subtitle_size}{Text size of subtitle in pt.}

\item{y_lab}{Y-axis label, use NULL for no label.}

\item{x_lab}{X-axis label, use NULL for no label.}

\item{background_color}{Color of the panel background.}

\item{panel_grid_color}{Color of the panel grid lines.}

\item{panel_grid_size}{Size of the panel grid lines in plot and contour lines around bars, useful to change if large dpi!}

\item{axis_size}{Size of the axis lines.}

\item{axis_text_angle}{Angle of the tick texts, 45 is recommended for many x levels.}

\item{text_size}{Size of the text in pt.}

\item{fill_colors}{Colors of the genders.}

\item{legend_pos}{Position of the legend in plot, if c(1,1), c(1,0) etc, legend inside plot.}

\item{legend_labels}{Label for each legend key.}

\item{label_breaks}{Order of the legend keys.}

\item{legend_row}{How many rows for the legends.}

\item{legend_col}{How many columns for the legends.}
}
\value{
Ggplot object containing age pyramid plot.
}
\description{
Plot an age pyramid using ggplot2.
}
\examples{
# Creating data
df <- data.frame(age = rpois(100000, 65), gender = sample(c('Woman', 'Woman', 'Man'), 100000, replace = TRUE))

# Age pyramid
age_pyramid(df = df, age_var = 'age', gender_var = 'gender', man_level = 'Man', title = "This is an age pyramid")

# Age pyramid with percent = FALSE
age_pyramid(df = df, age_var = 'age', gender_var = 'gender', man_level = 'Man', percent = FALSE, x_breaks = 5000, title = "This is an age pyramid")
}
