% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avvikelse.R
\name{avvikelse}
\alias{avvikelse}
\title{Calculate deviation (avvikelse) for each hospital using a linear model}
\usage{
avvikelse(prom_dat, out, preop = NULL, postop, decimals = 2)
}
\arguments{
\item{prom_dat}{PROM-data frame.}

\item{out}{Linear model outcome variable. Use string.}

\item{preop}{Preoperative variable. Use string.}

\item{postop}{Postoperative variable. Use string.}

\item{decimals}{Number of decimals.}
}
\value{
Data frame including hospitals, number of patients, preoperative
  values (if preop not NULL), postoperative values, and devation.
}
\description{
Calculates number of patients, preoperative values, postoperative values,
and devation for each hospital. Used to make some tables in the annual report.
}
\details{
The following variables MUST be included in the data set to be able to use
the function (used to calculate deviation):

"PREP_HipPain", "PREP_VASHealth", "PREP_EQ5D3Lindex",
"eq10", "eq20", "eq30", "eq40", "eq50"
"P_DiaGrp", "P_Gender", "P_Age", "PREP_Charnley"

where "eq10", "eq20", "eq30", "eq40" and "eq50" are the five numbers
extracted using \code{\link{eq5d_3l}}.
}
