% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compass.R
\name{compass}
\alias{compass}
\title{Compass plot function}
\usage{
compass(
  df,
  riket_name = "Riket",
  ncol = 6,
  riket_color = "#D95F02",
  hosp_color = "#1B9E77",
  alpha_riket = 0.6,
  alpha_hosp = 0.6,
  titles = NULL,
  title_size = 8,
  line_size = 0.1
)
}
\arguments{
\item{df}{Data frame.}

\item{riket_name}{Character to find what row is Swedish average, usually
named "Riket".}

\item{ncol}{Number of plots on each row.}

\item{riket_color}{Color of Swedish average polygon.}

\item{hosp_color}{Color of hospital polygon.}

\item{alpha_riket}{Modify color transparency for Swedish average color. Use <
1 if colors are to be mixed.}

\item{alpha_hosp}{Modify color transparency for hospital color. Use < 1 if
colors are to be mixed.}

\item{titles}{Character vector containing compass titles.}

\item{title_size}{Text size of title in pt.}

\item{line_size}{Size of the lines in the choord, useful to change if large
dpi!}
}
\value{
Ggplot object containing plot of compass or many compasses.
}
\description{
Radar plot using ggplot2.
}
\examples{
# Create 6 compasses with normalized data

compass(norm_compass_data[c(1:6,68),])

}
