% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compass_data.R
\docType{data}
\name{compass_data}
\alias{compass_data}
\title{Compass dataset}
\format{A data frame with 68 rows and 9 variables:
\describe{
  \item{Unit}{hospital name, ordered factor}
  \item{Satisfaction}{satisfaction mean, numeric}
  \item{Pain reduction}{pain reduction before operation and 1 year after, numeric}
  \item{Gain in EQ5D}{gain in EQ5D-index before operation and 1 year after, numeric}
  \item{AE90}{adverse events within 90 days, numeric}
  \item{Coverage}{coverage, numeric}
  \item{Reop 2 years}{reoperation within 2 years, numeric}
  \item{Implant 5 years}{implant survival within 5 years, numeric}
  \item{Implant 10 years}{implant survival within 10 years, numeric}
}}
\usage{
compass_data
}
\description{
A dataset containing hospital means for 9 different variables. This dataset was used in 2016 annual report.
Used for norm_compass() example, where outcome is used for compass() and compass_riket() examples.
}
\keyword{datasets}
