% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5dswe.r
\name{eq5d_3l}
\alias{eq5d_3l}
\title{Simple function to extract dimensions of EQ-5D profile}
\usage{
eq5d_3l(input, place)
}
\arguments{
\item{input}{EQ-5D profile, usually a five-digit number.}

\item{place}{Number from 1 to 5 indicating which number to extract.}
}
\value{
Number if input is a five-digit number, else NA.
}
\description{
Extracts a specific one-digit number from a five-digit number.
}
\examples{
eq5d_3l(input = 12312, 3)
# [1] 3

eq5d_3l(input = NA, 3)
# [1] NA

eq5d_3l(input = 1231, 3)
# [1] NA

eq5d_3l(input = c(12312, NA, 1231, 11121), 3)
# [1]  3 NA NA  1
}
