% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5dswe.r
\name{eqindBrit}
\alias{eqindBrit}
\title{Calculate Brittish EQ-5D-index using all five values extracted from eq5d_3l()}
\usage{
eqindBrit(Mobility, Selfcare, Usual, Pain, Anxiety)
}
\arguments{
\item{Mobility}{First number extracted from eq5d_3l().}

\item{Selfcare}{Second number extracted from eq5d_3l().}

\item{Usual}{Third number extracted from eq5d_3l().}

\item{Pain}{Fourth number extracted from eq5d_3l().}

\item{Anxiety}{Fifth number extracted from eq5d_3l().}
}
\value{
Number, EQ-5D-index.
}
\description{
Calculates Brittish EQ-ED-index using a known linear model (Dolan).
}
\examples{
# Returns intercept
eqindBrit(1,1,1,1,1)
# [1] 1

eqindBrit(NA,1,1,1,1)
# [1] NA

eqindBrit(5,1,1,1,1)
# [1] NA
}
