% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5dswe.r
\name{eqindSwed}
\alias{eqindSwed}
\title{Calculate Swedish EQ-5D-index using all five values extracted from eq5d_3l()}
\usage{
eqindSwed(Mobility, Selfcare, Usual, Pain, Anxiety)
}
\arguments{
\item{Mobility}{First number extracted from eq5d_3l().}

\item{Selfcare}{Second number extracted from eq5d_3l().}

\item{Usual}{Third number extracted from eq5d_3l().}

\item{Pain}{Fourth number extracted from eq5d_3l().}

\item{Anxiety}{Fifth number extracted from eq5d_3l().}
}
\value{
Number, EQ-5D-index.
}
\description{
Calculates Swedish EQ-ED-index using a known linear model (Model 4).
}
\examples{
# Returns intercept
eqindSwed(1,1,1,1,1)
# [1] 0.9694

eqindSwed(NA,1,1,1,1)
# [1] NA

eqindSwed(5,1,1,1,1)
# [1] NA
}
