% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected.R
\name{expected}
\alias{expected}
\title{Expected PROM}
\usage{
expected(outcome, data1, data2, data3, data4, year_period1 = "2008_2009",
  year_period2 = "2010_2011", year_period3 = "2012_2013",
  year_period4 = "2014_2015", mean_group = "Unit")
}
\arguments{
\item{outcome}{Linear model outcome variable.}

\item{data1}{Data frame containing data for the first year period.}

\item{data2}{Data frame containing data for the second year period.}

\item{data3}{Data frame containing data for the third year period.}

\item{data4}{Data frame containing data for the fourth year period.}

\item{year_period1}{First year period.}

\item{year_period2}{Second year period.}

\item{year_period3}{Third year period.}

\item{year_period4}{Fourth year period.}

\item{mean_group}{Group to take mean on.}
}
\value{
Data frame containing expected PROM for four year periods.
}
\description{
Expected PROM using linear model. Used to create expected data frame which is used in prom_trends().
}
\details{
Following variables MUST be included in the data set to be able to use the function:

PREP_HipPain

PREP_VASHealth

eq5d0swe

eq10

eq20

eq30

eq40

eq50

P_DiaGrp

P_Gender

P_Age

PREP_Charnley

where eq10, eq20, eq30, eq40 and eq50 are the five numbers extracted using eq5d_3l() and eq5d0swe is obtained using
eq10, eq20, eq30, eq40, and eq50 in eqindSwed().
}
