% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_compass.R
\name{norm_compass}
\alias{norm_compass}
\title{Normalization for compasses}
\usage{
norm_compass(x, riket_index = length(x))
}
\arguments{
\item{x}{Vector, commonly a column in a data frame.}

\item{riket_index}{Index where 'riket'-row exist.}
}
\value{
Normalized vector.
}
\description{
Function used to create data frame for compasses.
}
\examples{
# Normalize data

compass_data[,-1] <- apply(compass_data[,-1], 2, norm_compass)

# Make sure that the lowest value is the worst value in every column.
# If it is reversed, do like this:

compass_data$AE90 <- 1 - compass_data$AE90

}
