% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed.R
\name{observed}
\alias{observed}
\title{Observed PROM}
\usage{
observed(
  outcome,
  data1,
  data2,
  data3,
  data4,
  year_period1 = "2008_2009",
  year_period2 = "2010_2011",
  year_period3 = "2012_2013",
  year_period4 = "2014_2015",
  mean_group = "Unit"
)
}
\arguments{
\item{outcome}{PROM outcome variable.}

\item{data1, data2, data3, data4}{Data frames containing data for period 1/2/3/4}

\item{year_period1, year_period2, year_period3, year_period4}{Period 1/2/3/4.}

\item{mean_group}{Group to take mean on.}
}
\value{
Data frame containing observed PROM for four year periods.
}
\description{
Observed PROM using mean for each mean_group. Used to create observed data frame which is
used in \code{\link{prom_trends}}.
}
