% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_plot.R
\name{trend_plot}
\alias{trend_plot}
\title{Trend plot function}
\usage{
trend_plot(df, x_var, y_var, y_breaks = 5, y_lim = c(54.9, 65.1),
  x_breaks = 5, y_lab = "Procent kvinnor", x_lab = "år",
  title = NULL, subtitle = NULL, title_size = 14,
  subtitle_size = 12, line_color = "#377EB8", line_size = 1,
  point_size = 1, background_color = "moccasin",
  panel_grid_color = "grey", panel_grid_size = 0.3, axis_size = 0.3,
  axis_text_angle = 0, text_size = 8)
}
\arguments{
\item{df}{Data frame.}

\item{x_var}{X variable.}

\item{y_var}{Y variable.}

\item{y_breaks}{Length between each break on y-axis.}

\item{y_lim}{Limit on y-axis.}

\item{x_breaks}{Length between each break on x-axis.}

\item{y_lab}{Y-axis label, use NULL for no label.}

\item{x_lab}{X-axis label, use NULL for no label.}

\item{title}{Plot title, NULL for no title.}

\item{subtitle}{Small text under title, NULL for no subtitle.}

\item{title_size}{Text size of title in pt.}

\item{subtitle_size}{Text size of subtitle in pt.}

\item{line_color}{Color of the line.}

\item{line_size}{Size of the line.}

\item{point_size}{Size of the points.}

\item{background_color}{Color of the panel background.}

\item{panel_grid_color}{Color of the panel grid lines.}

\item{panel_grid_size}{Size of the panel grid lines in plot and contour
lines around bars, useful to change if large dpi!}

\item{axis_size}{Size of the axis lines.}

\item{axis_text_angle}{Angle of the tick texts, 45 is recommended for many x
levels.}

\item{text_size}{Size of the text in pt.}
}
\value{
Ggplot object containg trend plot.
}
\description{
Simple function to show trend in percent (y_var) between different years
(x_var) using geom_smooth (ggplot2).
}
\examples{

# Creating data

df <- data.frame(year = 2000:2017, prob = rnorm(18, 0.5, 0.02))

# Trend

trend_plot(df = df, x_var = 'year', y_var = 'prob', y_breaks = 2, y_lim = range(df$prob)*100)
}
